requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==5.2.7
django-stubs==5.2.7
django-stubs-ext==5.2.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
kombu==5.5.4
mysqlclient==2.2.7
openai==2.2.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
pydantic==2.12.0
pydantic_core==2.41.1
python-dateutil==2.9.0.post0
redis==6.4.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
types-PyYAML==6.0.12.20250915
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
